server/requirements.txt
-i https://pypi.org/simple
alembic==1.18.4; python_version >= '3.10'
aniso8601==10.0.1
bcrypt==5.0.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
click==8.3.1; python_version >= '3.10'
flask==3.1.3; python_version >= '3.9'
flask-bcrypt==1.0.1
flask-cors==6.0.2; python_version >= '3.9' and python_version < '4.0'
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
flask-migrate==4.1.0; python_version >= '3.6'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.3.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.3; python_version >= '3.9'
pyjwt==2.11.0; python_version >= '3.9'
pytz==2026.1.post1
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.48; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
werkzeug==3.1.6; python_version >= '3.9'
psycopg2-binary==2.9.9
python-dotenv==1.0.0